+++ main.py
@@ -1,6 +1,6 @@
 """
 Sinote: Code Smartly, User Interface Intuitive
-Sinote Version 0.07.26031
+Sinote Version 0.07.26033
 3rd Party License saved at ./3rd-party-license
 
 Win12Home (C) 2025, 2026. MIT License.