+++ utils/const.py
@@ -1,3 +1,3 @@
 apiVersion: tuple[int, int, int] = (1, 0, 2)
-sinoteVersion: str = "sinote-0.1-initial-preview-beta-0.07.26031"
-__version__: tuple[str, str, str] = ("0", "07", "26031")
+sinoteVersion: str = "sinote-0.1-initial-preview-beta-0.08.26033"
+__version__: tuple[str, str, str] = ("0", "08", "26033")